requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
psycopg2-binary
python-dotenv==1.0.0
fpdf2
python-docx==1.1.0
gunicorn==21.2.0
requests==2.31.0
huggingface_hub
